feat(errorControl): 값이 없을때 표시 추가
검색 결과가 없거나 데이터가 없으면 없다고 따로 표현하는 로직 추가.
