+++ src/main.tsx
@@ -17,10 +17,10 @@ function renderApp() {
     </React.StrictMode>,
   );
 }
-if (process.env.NODE_ENV === 'development') {
-  worker.start().then(() => {
-    renderApp();
-  });
-} else {
-  renderApp();
-}
+// if (process.env.NODE_ENV === 'development') {
+//   worker.start().then(() => {
+//     renderApp();
+//   });
+// } else {
+renderApp();
+//}
